Fixed the deflate-code for compilation on 64-bit

+++ tools/elf2e32/deflate/deflate/inflate.cpp
@@ -48,7 +48,7 @@ TInt CInflater::ReadL(TUint8* aBuffer,TInt aLength)
 	TInt tfr=0;
 	for (;;)
 		{
-		TInt len=Min(aLength,iLimit-iAvail);
+		TInt len=Min((intptr_t)aLength,iLimit-iAvail);
 		if (len && aBuffer)
 			{
 			HMem::Copy(aBuffer,iAvail,len);
@@ -137,7 +137,7 @@ TInt CInflater::InflateL()
 			iRptr+=KDeflateMaxDistance;
 		}
 useHistory:
-		TInt tfr=Min(end-out,iLen);
+		TInt tfr=Min(end-out,(intptr_t)iLen);
 		iLen-=tfr;
 		const TUint8* from=iRptr;
 		do
